remove IGItems and IGBlocks classes
